+++ lib/rig_api/v1/responses.ex
@@ -5,7 +5,8 @@ defmodule RigApi.V1.Responses do
   use RigApi, :controller
   use PhoenixSwagger
 
-  alias RigInboundGateway.ApiProxy.ResponseFromParser
+  alias Rig.Connection.Codec
+  alias RigCloudEvents.CloudEvent
 
   @prefix "/v1"
 
@@ -32,19 +33,20 @@ defmodule RigApi.V1.Responses do
 
   @doc """
   Accepts message to be sent to correlated HTTP process.
-
   Note that body has to contain following field `"rig": { "correlation": "_id_" }`.
   """
-  def create(%{req_headers: req_headers} = conn, message) do
-    case ResponseFromParser.parse(req_headers, message) do
-      {deserialized_pid, response_code, response} ->
-        Logger.debug(fn ->
-          "HTTP response via internal HTTP to #{inspect(deserialized_pid)}: #{inspect(message)}"
-        end)
+  def create(conn, message) do
+    with {:ok, cloud_event} <- CloudEvent.parse(message),
+         {:ok, rig_metadata} <- Map.fetch(message, "rig"),
+         {:ok, correlation_id} <- Map.fetch(rig_metadata, "correlation"),
+         {:ok, deserialized_pid} <- Codec.deserialize(correlation_id) do
+      Logger.debug(fn ->
+        "HTTP response via internal HTTP to #{inspect(deserialized_pid)}: #{inspect(message)}"
+      end)
 
-        send(deserialized_pid, {:response_received, response, response_code})
-        send_resp(conn, :accepted, "message sent to correlated reverse proxy request")
-
+      send(deserialized_pid, {:response_received, cloud_event.json})
+      send_resp(conn, :accepted, "message sent to correlated reverse proxy request")
+    else
       err ->
         Logger.warn(fn -> "Parse error #{inspect(err)} for #{inspect(message)}" end)