+++ test/rig_tests/proxy/response_from/async_http_test.exs
@@ -1,6 +1,6 @@
 defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
   @moduledoc """
-  If `response_from` is set to http_async, the response is taken from internal HTTP endpoint /v2/responses
+  If `response_from` is set to http_async, the response is taken from internal HTTP endpoint /v3/responses
 
   Note that `test_with_server` sets up an HTTP server mock, which is then configured
   using the `route` macro.
@@ -41,74 +41,13 @@ defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
       |> put_req_header("rig-correlation", correlation_id)
       |> put_req_header("rig-response-code", "201")
       |> put_req_header("content-type", "application/json;charset=utf-8")
-      |> post("/v2/responses", Jason.encode!(async_response))
+      |> post("/v3/responses", Jason.encode!(async_response))
 
       Response.ok!(sync_response, %{"content-type" => "application/json"})
     end)
 
     # We register the endpoint with the proxy:
-    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
-    rig_proxy_url = "http://localhost:#{@proxy_port}"
-
-    body =
-      Jason.encode!(%{
-        id: api_id,
-        name: "Mock API",
-        version_data: %{
-          default: %{
-            endpoints: [
-              %{
-                id: endpoint_id,
-                method: "GET",
-                path: endpoint_path,
-                response_from: "http_async"
-              }
-            ]
-          }
-        },
-        proxy: %{
-          target_url: "localhost",
-          port: FakeServer.port()
-        }
-      })
-
-    headers = [{"content-type", "application/json"}]
-    HTTPoison.post!(rig_api_url, body, headers)
-
-    # The client calls the proxy endpoint:
-    request_url = rig_proxy_url <> endpoint_path
-    %HTTPoison.Response{status_code: res_status, body: res_body} = HTTPoison.get!(request_url)
-
-    # Now we can assert that...
-    # ...the fake backend service has been called:
-    assert FakeServer.hits() == 1
-    # ...the connection is closed and the status is OK:
-    assert res_status == 201
-    # ...but the client got the response sent to the HTTP internal endpoint:
-    assert Jason.decode!(res_body) == async_response
-  end
-
-  test_with_server "Given (v1) response_from is set to http_async and response is in binary mode, the http response  is taken from the internal HTTP endpoint." do
-    test_name = "proxy-http-response-from-http-internal-binary-v1"
-
-    api_id = "mock-#{test_name}-api"
-    endpoint_id = "mock-#{test_name}-endpoint"
-    endpoint_path = "/#{endpoint_id}"
-    sync_response = %{"this response" => "the client never sees this response"}
-    async_response = %{"message" => "this is the async response that reaches the client instead"}
-
-    route(endpoint_path, fn %{query: %{"correlation" => correlation_id}} ->
-      build_conn()
-      |> put_req_header("rig-correlation", correlation_id)
-      |> put_req_header("rig-response-code", "201")
-      |> put_req_header("content-type", "application/json;charset=utf-8")
-      |> post("/v1/responses", Jason.encode!(async_response))
-
-      Response.ok!(sync_response, %{"content-type" => "application/json"})
-    end)
-
-    # We register the endpoint with the proxy:
-    rig_api_url = "http://localhost:#{@api_port}/v1/apis"
+    rig_api_url = "http://localhost:#{@api_port}/v3/apis"
     rig_proxy_url = "http://localhost:#{@proxy_port}"
 
     body =